Enable cycling support for Gaea C6 (#3323)
Add support for the Gaea clusters by enabling cycled experiments
on C6 and addressing a number of issues on both clusters.

Also included is support for HPSS on C6, but it should NOT be
utilized at this time. The ES cluster, where HPSS connections are made,
has a significant issue with the F6 (C6's filesystem) mount that causes
the system to run extremely slow when filesystem-intense operations are
performed (such as htar). There is a plan to enable this feature more
broadly in the near future.

Included changes:
- Fixed memory variable unsetting for Gaea C5/6 in config.resources.GAEAC{5,6}
- Refactoring the system-level parameter detection when determining task
resources in the setup scripts to make it easier to define multiple
partitions, queues, and clusters.
- Adding a `DTN` partition, queue, and cluster definition.
- Added/renamed missing/miss-named tasks to tasks.py and added a check
that the input task is valid.

Refs #2905
Refs #3108
Refs #3133 
Refs #3261
Refs #3269 
Refs #3324