Adding Jenkins CI to GaeaC6 using role account (#3389)
# Description

This PR has the necessary updates to the CI scripts in Jenkins for
running on **Gaea C6**.
The **GaeaC6-EMC** Node is all ready configured and connected to the
**Gaea C6** Head Node `gaea64`

[PyGithub](https://github.com/PyGithub/PyGithub) is not in **Spack
Stack** yet and is needed in support of the GitHub logging and gist
creation python codes. This python module will be installed in
`$HOME/.local` in the role accounts until the module is added to **Spack
Stack**.

Same with **GitHub CLI** (which will be supported in Spack Stack 1.9) it
will be installed in `${ROLE_HOME}/utils/bin` which should be in role
account's `$PATH `where `ROLE_HOME` is at the same level as `fix` and
`data`.

This is setup to run in the role account with the following added
directories and utils:

`/gpfs/f6/drsa-precip3/proj-shared/global/CI`

Location on disk where the CI test cases actually build and run
specified in
[Jenkinsfile](https://github.com/TerrenceMcGuinness-NOAA/global-workflow/blob/02486bcaf76ee9dbbffd07e0932fe4edb98f7a34/ci/Jenkinsfile#L9)

```
└── GFS_CI_CD
    ├── global-workflow
    │   └── workflow
    ├── Bash
    ├── GitLab
    └── Jenkins
        ├── agent
        └── workspace

```
Top level CI directories (this will file structure be duplicated for
each role account)
```
GFS_CI_CD
```
CI system directory for Jenkins launching scripts and in support of BASH
CI
```
GFS_CI_CD/global-workflow
```
Directory where the JAVA agent is launched from
```
Jenkins/agent
```
Jenkins work space (system level workspace)
```
Jenkins/workspace
```

**GitHub CLI (`gh` utility):**

Location where `gh` was installed (pending when supporting Spack Stack
1.9)
```
gpfs/f6/drsa-precip3/world-shared/role.glopara/utils/bin/gh
```
soft link to gh utility for Jenkins to find
```
/ncrc/home1/role.glopara/bin
```
  Resolves #3375

# Type of change
- [ ] Bug fix (fixes something broken)
- [ ] New feature (adds functionality)
- [x] Maintenance (code refactor, clean-up, new CI test, etc.)

# Change characteristics
<!-- Choose YES or NO from each of the following and delete the other
-->
- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? NO
- Does this change require an update to any of the following submodules?
NO (If YES, please add a link to any PRs that are pending.)
  - [ ] EMC verif-global <!-- NOAA-EMC/EMC_verif-global#1234 -->
  - [ ] GDAS <!-- NOAA-EMC/GDASApp#1234 -->
  - [ ] GFS-utils <!-- NOAA-EMC/gfs-utils#1234 -->
  - [ ] GSI <!-- NOAA-EMC/GSI#1234 -->
  - [ ] GSI-monitor <!-- NOAA-EMC/GSI-Monitor#1234 -->
  - [ ] GSI-utils <!-- NOAA-EMC/GSI-Utils#1234 -->
  - [ ] UFS-utils <!-- ufs-community/UFS_UTILS#1234 -->
  - [ ] UFS-weather-model <!-- ufs-community/ufs-weather-model#1234 -->
  - [ ] wxflow <!-- NOAA-EMC/wxflow#1234 -->

# How has this been tested?

This PR can be tested in place after Dave's PR is merged for cycling on
GaeaC6

---------

Co-authored-by: Terry McGUinness <terrene.mcguinness@gmail.com>
Co-authored-by: Terry McGuinness <terry.mcguinness@noaa.gov>
Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>
Co-authored-by: emc.glopara <emc.glopara@noaa.gov>
Co-authored-by: David Huber <69919478+DavidHuber-NOAA@users.noreply.github.com>